package setup and deployment
Prepared to be installed and deployed with pip
