+++ MANIFEST
@@ -1,0 +1,2 @@
+# file GENERATED by distutils, do NOT edit
+setup.py
