+++ keeper/__init__.py
@@ -1,0 +1,11 @@
+import sys
+
+from keeper.cli import loop
+
+
+def main():
+    loop()
+
+
+if __name__ == "__main__":
+    sys.exit(main())
