+++ keeper/plugin_manager.py
@@ -9,7 +9,6 @@
 # Contact: ops@keepersecurity.com
 #
 
-import sys
 import importlib
 
 imported_plugins = {}
