+++ keeper/plugins/.gitignore
