+++ keeper/plugins/adpasswd/adpasswd.py
