--- src/config.json.sample
@@ -1,10 +1,0 @@
-{
-    "server":"https://keeperapp.com/v2/",
-    "email":"",
-    "password":"",
-    "mfa_token":"",
-    "mfa_type":"",
-    "debug":false,
-    "plugins":[],
-    "commands":[]
-}
